Move canned policy support into CloudFront signer; remove custom policy generation
